extension: util: Move sm_log to util
Since we'd like to use it from other modules.
